+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `modal` high-order component.
+- Feature: Add `Modal` component.
 - Feature: Add `SimpleUploader`, `Uploader` and `ImageCropper` components.
 - Fix: Handles small version of `InfoLines` component.
 - Fix: Harmonize font-size label component on `FormInfo`.